+++ app/src/main/java/com/example/administrator/ccoupons/Fragments/User/UserOptionFragment.java
@@ -57,7 +57,9 @@ public class UserOptionFragment extends Fragment implements AppBarLayout.OnOffse
     @BindView(R.id.user_portrait)
     CircleImageView portrait;
 
-    @OnClick({R.id.user_main_toolbar, R.id.user_nickname_text, R.id.user_to_mycoupons, R.id.user_to_wal, R.id.user_to_inf, R.id.user_sell, R.id.user_buy, R.id.user_follow, R.id.user_to_set, R.id.user_logoff, R.id.user_portrait})
+    @OnClick({R.id.user_main_toolbar, R.id.user_nickname_text, R.id.user_to_mycoupons,
+            R.id.user_to_wal, R.id.user_to_inf, R.id.user_sell, R.id.user_buy,
+            R.id.user_follow, R.id.user_to_set, R.id.user_logoff, R.id.user_portrait, R.id.user_to_about})
     public void onViewClicked(View view) {
         switch (view.getId()) {
             case R.id.user_main_toolbar:
@@ -90,8 +92,24 @@ public class UserOptionFragment extends Fragment implements AppBarLayout.OnOffse
             case R.id.user_portrait:
                 getActivity().startActivity(new Intent(getActivity(), UserInformationActivity.class));
                 break;
+            case R.id.user_to_about:
+                LayoutInflater inflater = getActivity().getLayoutInflater();
+                View dialog = inflater.inflate(R.layout.help_layout, (ViewGroup) getActivity().findViewById(R.id.help_root));
+                AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+                builder.setTitle("U惠 Release 2.0");
+                builder.setPositiveButton("确定", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.dismiss();
+                    }
+                });
+                builder.setView(dialog);
+                builder.setIcon(R.drawable.welcome_logo_new);
+                builder.show();
+                break;
         }
     }
+
     Unbinder unbinder;
     private UserInfoManager userInfoManager;
     private static final float PERCENTAGE_TO_SHOW_TITLE_AT_TOOLBAR = 0.6f;
