+++ app/src/main/java/com/example/administrator/ccoupons/User/UserSettingActivity.java
@@ -1,5 +1,6 @@
 package com.example.administrator.ccoupons.User;
 
+import android.content.Context;
 import android.content.DialogInterface;
 import android.support.v7.app.AlertDialog;
 import android.os.Bundle;
@@ -15,6 +16,7 @@ import com.example.administrator.ccoupons.Tools.SlideBackActivity;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
+import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
 public class UserSettingActivity extends SlideBackActivity {
