+++ app/src/main/java/com/example/administrator/ccoupons/User/UserWalletActivity.java
@@ -1,5 +1,6 @@
 package com.example.administrator.ccoupons.User;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
@@ -12,8 +13,15 @@ import com.example.administrator.ccoupons.Tools.SlideBackActivity;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
+import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
 public class UserWalletActivity extends SlideBackActivity {
+
+    @Override
+    protected void attachBaseContext(Context newBase) {
+        super.attachBaseContext(CalligraphyContextWrapper.wrap(newBase));
+    }
+
     @BindView(R.id.uwal_toolbar)
     Toolbar toolbar;
     @BindView(R.id.uwal_ub)
@@ -61,8 +69,7 @@ public class UserWalletActivity extends SlideBackActivity {
      * init Ucoin data
      */
     private void initData() {
-        ub = (TextView) findViewById(R.id.uwal_ub);
         MyApp app = (MyApp) getApplicationContext();
-        ub.setText(app.getUcoin() + "");
+        ub.setText(String.format("%.2f", (float)app.getUcoin()));
     }
 }
