+++ app/src/main/res/layout/activity_user_wallet.xml
@@ -37,42 +37,50 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/colorPrimary"
-        android:orientation="horizontal">
+        android:orientation="vertical">
 
         <TextView
             style="@style/user_text"
             android:text="@string/uwal_ub"
             android:textColor="@color/white"
-            android:textSize="35dp" />
+            android:textSize="15dp" />
 
-        <TextView
-            android:id="@+id/uwal_ub"
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
 
-            style="@style/user_text"
-            android:layout_marginRight="10dp"
-            android:gravity="right"
-            android:textColor="@color/white"
-            android:textSize="40dp" />
+            <TextView
+                android:id="@+id/uwal_ub"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:padding="10dp"
+                android:layout_marginRight="10dp"
+                android:gravity="left"
+                android:textColor="@color/white"
+                android:textSize="40dp" />
 
-        <ImageView
-            style="@style/user_icon_layout"
-            android:layout_width="30dp"
-            android:layout_height="30dp"
-            android:layout_gravity="center"
-            android:src="@drawable/plus" />
+            <ImageView
+                style="@style/user_icon_layout"
+                android:layout_width="40dp"
+                android:layout_height="40dp"
+                android:layout_gravity="center"
+                android:src="@drawable/plus" />
+        </LinearLayout>
     </LinearLayout>
+
 
     <Button
         android:id="@+id/button_purchase_ub"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
+        android:layout_marginTop="30dp"
         android:background="@color/colorPrimary"
         android:paddingBottom="5dp"
         android:paddingLeft="60dp"
         android:paddingRight="60dp"
         android:paddingTop="5dp"
-        android:layout_marginTop="30dp"
         android:text="充值U币"
         android:textColor="@color/white"
         android:textSize="25dp" />
